+++ packages/data-types/src/lib/interfaces/fruit-plant.interface.ts
@@ -1,0 +1,5 @@
+import { BaseCrop } from "./base-crop.interface";
+
+export interface FruitPlant extends BaseCrop {
+
+}