+++ packages/guide/src/app/shared/components/processing-time-per-quality/processing-time-per-quality.component.ts
@@ -1,0 +1,18 @@
+import { Component, Input } from '@angular/core';
+import { Quality } from "@ci/data-types";
+import { UiIcon } from "../../enums/ui-icon.enum";
+
+@Component({
+    selector: 'app-processing-time-per-quality',
+    templateUrl: './processing-time-per-quality.component.html',
+    styleUrls: ['./processing-time-per-quality.component.scss'],
+})
+export class ProcessingTimePerQualityComponent {
+
+    @Input() baseProcessingTime?: { day: number; time: { hours: number, minutes: number } };
+    @Input() processingTimePerQualities?: Record<string | Quality, { day: number; time: { hours: number, minutes: number } }>
+
+    protected qualityValues: Quality[] = Object.values(Quality);
+    protected quality = Quality;
+    protected uiIcon = UiIcon;
+}