+++ packages/guide/src/app/shared/pipes/spaces-to-pascal-case.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { SpacesToPascalCasePipe } from './spaces-to-pascal-case.pipe';
+
+describe('SpacesToPascalCasePipe', () => {
+    it('create an instance', () => {
+        const pipe = new SpacesToPascalCasePipe();
+        expect(pipe).toBeTruthy();
+    });
+});