+++ packages/pak-files-parser/src/interfaces/raw-unlock-by-mastery.interface.ts
@@ -1,4 +1,4 @@
-export interface RawCraftingUnlockByMastery {
+export interface RawUnlockByMastery {
     "masteryLevel": number,
     "desc": string,
     "unlockRecipe": {