+++ packages/pak-files-parser/src/main.ts
@@ -16,6 +16,10 @@ import { GiftPreferencesDbGenerator } from './app/gift-preferences-db.generator'
 import { NPCDbGenerator } from './app/npc-db.generator';
 import { CraftingRecipeUnlockedByMasteryDbGenerator } from "./app/crafting-recipe-unlocked-by-mastery-db.generator";
 import { config } from "./config";
+import { CookingRecipeUnlockedByMasteryDbGenerator } from "./app/cooking-recipe-unlocked-by-mastery-db.generator";
+import { CookingDbGenerator } from "./app/cooking-db.generator";
+import { FruitTreeDbGenerator } from "./app/fruit-tree-db.generator";
+import { FruitPlantDbGenerator } from "./app/fruit-plant-db.generator";
 
 const itemIconPath = config.itemIconPath
 const texturePath = config.texturePath;
@@ -29,14 +33,22 @@ const npcDbMap = npcDbGenerator.generate();
 const craftingRecipeUnlockedByMasteryDbGenerator = new CraftingRecipeUnlockedByMasteryDbGenerator(itemDbMap);
 const craftingRecipeUnlockedByMasteryDbMap = craftingRecipeUnlockedByMasteryDbGenerator.generate();
 
+const cookingRecipeUnlockedByMasteryDbGenerator = new CookingRecipeUnlockedByMasteryDbGenerator(itemDbMap);
+const cookingRecipeUnlockedByMasteryDbMap = cookingRecipeUnlockedByMasteryDbGenerator.generate();
+
+const tagBasedItemsDbGenerator = new TagBasedItemGenericDbGenerator();
+const tagBasedItemsDbMap = tagBasedItemsDbGenerator.generate();
+
 const generators: Record<string, { generate: () => Map<string, any> }> = {
     'crafting-recipes': new CraftingRecipeDbGenerator(itemDbMap, craftingRecipeUnlockedByMasteryDbMap),
     'bugs-and-insects': new BugsAndInsectsDbGenerator(itemDbMap),
     'ocean-critters': new OceanCritterDbGenerator(itemDbMap),
     'fish': new FishDbGenerator(itemDbMap),
     'crops': new CropsDbGenerator(itemDbMap),
-    'tag-based-items': new TagBasedItemGenericDbGenerator(),
+    'fruit-trees': new FruitTreeDbGenerator(itemDbMap),
+    'fruit-plants': new FruitPlantDbGenerator(itemDbMap),
     'item-processing': new ItemProcessorDbGenerator(itemDbMap),
+    'cooking-recipes': new CookingDbGenerator(itemDbMap, cookingRecipeUnlockedByMasteryDbMap, tagBasedItemsDbMap),
 
     'journal-fish': new JournalOrderDbGenerator('Caught/DT_JournalFish.json'),
     'journal-insects': new JournalOrderDbGenerator('Caught/DT_JournalInsects.json'),
@@ -55,7 +67,10 @@ const generators: Record<string, { generate: () => Map<string, any> }> = {
     'gift-preferences': new GiftPreferencesDbGenerator(itemDbMap, npcDbMap),
     'npcs': new NPCDbGenerator(itemDbMap),
 
+
+    'tag-based-items': {generate: () => tagBasedItemsDbMap},
     'crafting-unlocks-by-mastery': {generate: () => craftingRecipeUnlockedByMasteryDbMap},
+    'cooking-unlocks-by-mastery': {generate: () => cookingRecipeUnlockedByMasteryDbMap},
 
     // last so applied changed will be written as well
     items: {generate: () => itemDbMap},